requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.4
numpy==1.24.3
matplotlib==3.8.2
seaborn==0.13.0
scikit-learn==1.3.2
yfinance==0.2.28
akshare==1.12.0
tushare==1.2.89
jieba==0.42.1
wordcloud==1.9.2
openpyxl==3.1.2
xlrd==2.0.1
selenium==4.15.2
webdriver-manager==4.0.1
fake-useragent==1.4.0
lxml==4.9.3
html5lib==1.1
urllib3==2.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6
soupsieve==2.5
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
kiwisolver==1.4.5
pillow==10.1.0
pyparsing==3.1.1
cycler==0.12.1
contourpy==1.2.0
fonttools==4.46.0
packaging==23.2
threadpoolctl==3.2.0
joblib==1.3.2
scipy==1.11.4
multiprocessing-logging==0.3.4
talib-binary==0.4.26
